:lipstick: feat: Arquivo de estilos para App
