+++ src/App.tsx
@@ -1,8 +1,8 @@
 import { useState } from "react";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
-import styled from "styled-components";
 import Header from "./components/Header";
 import Sidebar from "./components/Sidebar";
+import { RightSide } from "./components/style";
 import Home from "./pages/Home";
 import GlobalStyle from "./styles/GlobalStyle";
 
@@ -26,21 +26,5 @@ const App = () => {
   );
 };
 
-interface Props {
-  sidebarIsActive: boolean;
-}
 
-const RightSide = styled.div<Props>`
-  position: fixed;
-  left: ${(props) => (props.sidebarIsActive ? "235px" : "0")};
-  width: ${(props) =>
-    props.sidebarIsActive ? "calc(100vw - 235px)" : "100vw"};
-  transition-timing-function: linear;
-  transition-duration: 0.5s;
-
-  @media (max-width: 1100px) {
-    left: 0;
-    width: 100vw;
-  }
-`;
 export default App;
